test_SaidiYassine/backend/filmrate/serializers.py
from rest_framework import serializers
from .models import Actor, Movie, Review
from django.db.models import Avg

class ActorSerializer(serializers.ModelSerializer):
    class Meta:
        model = Actor
        fields = '__all__'

class ReviewSerializer(serializers.ModelSerializer):
    class Meta:
        model = Review
        fields = '__all__'

class MovieListSerializer(serializers.ModelSerializer):
    class Meta:
        model = Movie
        fields = ['id', 'title']

class MovieDetailSerializer(serializers.ModelSerializer):
    actors = ActorSerializer(many=True, read_only=True)
    average_grade = serializers.SerializerMethodField()

    class Meta:
        model = Movie
        fields = '__all__'

    def get_average_grade(self, obj):
        average = obj.reviews.aggregate(Avg('grade'))['grade__avg']
        return round(average, 2) if average is not None else None
    
